[FEAT] Add `register` function
This should ease component registration. Components passed need a `type`
property, and two optional props: `property`, and `topology`.

Additionally, the component is wrapped automatically with the `link`
function when a static `mapDataToProps` property is present on the
component.

The requirement for `type` may be dropped when we're able to infer the
type from the prop map.